Fix parametrisation of logs to keep.

+++ src/gocept/jenkinsdsl/templates/builder.abstract.groovy
@@ -41,7 +41,7 @@ class AbstractBuilder implements Builder {
             }
 
             // logRotator(daysToKeepInt=100, numToKeepInt=100) did not work in 1.43
-            logRotator(this.log_days, this.log_builds)
+            logRotator(this.log_days as int, this.log_builds as int)
             checkoutRetryCount()
         }
     }
